requirements.txt
certifi
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.8
django-allauth==0.34.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-rest-auth==0.9.2
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
pip==9.0.1
Pygments==2.2.0
pypinyin==0.28.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
setuptools==38.4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
wheel==0.30.0
selenium==3.11.0
